chore: move linking code to forge crate
