+++ crates/forge/bin/cmd/script/build.rs
@@ -1,6 +1,7 @@
 use super::*;
 use alloy_primitives::{Address, Bytes};
 use eyre::{Context, ContextCompat, Result};
+use forge::link::{link_with_nonce_or_address, PostLinkInput, ResolvedDependency};
 use foundry_cli::utils::get_cached_entry_by_name;
 use foundry_common::{
     compact_to_contract,
@@ -14,7 +15,6 @@ use foundry_compilers::{
     info::ContractInfo,
     ArtifactId, Project, ProjectCompileOutput,
 };
-use foundry_utils::{PostLinkInput, ResolvedDependency};
 use std::{collections::BTreeMap, str::FromStr};
 
 impl ScriptArgs {
@@ -119,7 +119,7 @@ impl ScriptArgs {
             }
         }
 
-        foundry_utils::link_with_nonce_or_address(
+        link_with_nonce_or_address(
             contracts.clone(),
             &mut highlevel_known_contracts,
             libs,
